+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-...
+- *psb* Continuous Strobe support for USB650 **thanks @moritzj29**
 
 ## [1.0.2] - 2020-05-04
 ### Fixed